+++ packages/amplify-category-custom/src/__tests__/walkthroughts/cloudformation-walkthrough.test.ts
@@ -8,7 +8,7 @@ jest.mock('../../utils/build-custom-resources');
 jest.mock('../../utils/dependency-management-utils');
 
 jest.mock('amplify-cli-core');
-jest.mock('amplify-prompts');
+jest.mock('@aws-amplify/amplify-prompts');
 
 jest.mock('fs-extra', () => ({
   readFileSync: jest.fn().mockReturnValue('mockCode'),