+++ packages/amplify-e2e-core/src/init/non-interactive-init.ts
@@ -1,4 +1,4 @@
-import execa from 'execa';
+import execa, { ExecaReturnValue } from 'execa';
 // eslint-disable-next-line import/no-cycle
 import { getCLIPath, TEST_PROFILE_NAME } from '..';
 import { CategoriesConfig, AwsProviderConfig } from './headless-types';
@@ -9,8 +9,8 @@ import { CategoriesConfig, AwsProviderConfig } from './headless-types';
 export const nonInteractiveInitAttach = async (
   projRoot: string,
   amplifyInitConfig: AmplifyInitConfig,
+  awsProviderConfig: AwsProviderConfig | AwsProviderGeneralConfig,
   categoriesConfig?: CategoriesConfig,
-  awsProviderConfig = getAwsProviderConfig(),
 ): Promise<void> => {
   const args = [
     'init',
@@ -37,7 +37,8 @@ export const nonInteractiveInitWithForcePushAttach = async (
   categoriesConfig?: CategoriesConfig,
   testingWithLatestCodebase = false,
   awsProviderConfig = getAwsProviderConfig(),
-): Promise<void> => {
+  rejectOnFailure = true,
+): Promise<ExecaReturnValue<string>> => {
   const args = [
     'init',
     '--yes',
@@ -48,11 +49,12 @@ export const nonInteractiveInitWithForcePushAttach = async (
       awscloudformation: awsProviderConfig,
     }),
     '--forcePush',
+    '--debug',
   ];
   if (categoriesConfig) {
     args.push('--categories', JSON.stringify(categoriesConfig));
   }
-  await execa(getCLIPath(testingWithLatestCodebase), args, { cwd: projRoot });
+  return execa(getCLIPath(testingWithLatestCodebase), args, { cwd: projRoot, reject: rejectOnFailure });
 };
 
 /**
@@ -67,10 +69,23 @@ export const getAmplifyInitConfig = (projectName: string, envName: string): Ampl
 /**
  * Returns a default AwsProviderConfig
  */
-export const getAwsProviderConfig = (): AwsProviderConfig => ({
-  configLevel: 'project',
-  useProfile: true,
-  profileName: TEST_PROFILE_NAME,
+export const getAwsProviderConfig = (profileType?: string): AwsProviderConfig | AwsProviderGeneralConfig => {
+  if (profileType === 'general') {
+    return getAwsProviderGeneralConfig();
+  } else {
+    return {
+      configLevel: 'project',
+      useProfile: true,
+      profileName: TEST_PROFILE_NAME,
+    };
+  }
+};
+
+/**
+ * Returns a general AwsProviderConfig
+ */
+export const getAwsProviderGeneralConfig = (): AwsProviderGeneralConfig => ({
+  configLevel: 'general',
 });
 
 /**
@@ -81,4 +96,8 @@ export type AmplifyInitConfig = {
   envName: string;
   defaultEditor: string;
   frontend?: string;
+};
+
+export type AwsProviderGeneralConfig = {
+  configLevel: string;
 };