+++ packages/amplify-cli-core/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-cli-core@4.0.1-rc.e9e0d157d.0...amplify-cli-core@4.0.1) (2023-03-23)
+
+**Note:** Version bump only for package amplify-cli-core
+
+
+
+
+
+## [4.0.1-rc.e9e0d157d.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-cli-core@4.0.0...amplify-cli-core@4.0.1-rc.e9e0d157d.0) (2023-03-17)
+
+
+### Bug Fixes
+
+* **amplify-console-hosting:** manual publish build command error output ([#12166](https://github.com/aws-amplify/amplify-cli/issues/12166)) ([43929bf](https://github.com/aws-amplify/amplify-cli/commit/43929bf7a136bf0fd68c809cfc8e9144544dde47))
+* cycle detection in yarn 1 and npm detector ([#12255](https://github.com/aws-amplify/amplify-cli/issues/12255)) ([aa07944](https://github.com/aws-amplify/amplify-cli/commit/aa07944d9972748c73c236974e138f08793ecfee))
+* invalid method call in interactions walkthrough ([#12108](https://github.com/aws-amplify/amplify-cli/issues/12108)) ([2d3b3b6](https://github.com/aws-amplify/amplify-cli/commit/2d3b3b6f6a343a5895acf4c51b26f7104e0fc794))
+* lint warnings ([#12180](https://github.com/aws-amplify/amplify-cli/issues/12180)) ([83d9bf3](https://github.com/aws-amplify/amplify-cli/commit/83d9bf35fa709084605c23bd2f54feadb3bdbb87))
+
+
+
+
+
 # [4.0.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-cli-core@3.7.1...amplify-cli-core@4.0.0) (2023-03-13)